refactor : useSpotlight, useTeamVoteResult, useTutorial을 features/battle/hooks로 이동
