+++ frontend/src/features/battle/hooks/useSpotlight.ts
