+++ frontend/src/features/battle/hooks/useTeamVoteResult.ts
